+++ flow/resource_mac_bare_metal_security_group_rule_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 )
 
 func TestAccMacBareMetalSecurityGroupRule_Basic(t *testing.T) {
@@ -33,6 +34,24 @@ func TestAccMacBareMetalSecurityGroupRule_Basic(t *testing.T) {
 					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "port_range.to", fmt.Sprint(toPort)),
 					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "ip_range", ipRange),
 					resource.TestCheckNoResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "icmp"),
+				),
+			},
+			{
+				// mac bare metal has no check on the pair, so a dropped code 0 reaches the
+				// vendor api as null and the rule comes back wider than asked for: the values
+				// read back are the assertion, not the absence of an error
+				Config: fmt.Sprintf(testAccMacBareMetalSecurityGroupRuleConfigICMP, securityGroupName, 8, 0, ipRange),
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction("flow_mac_bare_metal_security_group_rule.foobar", plancheck.ResourceActionUpdate),
+					},
+				},
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "protocol.number", "1"),
+					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "protocol.name", "icmp"),
+					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "icmp.type", "8"),
+					resource.TestCheckResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "icmp.code", "0"),
+					resource.TestCheckNoResourceAttr("flow_mac_bare_metal_security_group_rule.foobar", "port_range"),
 				),
 			},
 		},
@@ -67,5 +86,35 @@ resource "flow_mac_bare_metal_security_group_rule" "foobar" {
 	}
 
 	ip_range = "%[5]s"
+}
+`
+
+const testAccMacBareMetalSecurityGroupRuleConfigICMP = `
+data "flow_location" "zrh1" {
+	name = "ZRH1"
+}
+
+resource "flow_mac_bare_metal_network" "foobar" {
+	name        = "%[1]s"
+	location_id = data.flow_location.zrh1.id
+}
+
+resource "flow_mac_bare_metal_security_group" "foobar" {
+	name       = "%[1]s"
+	network_id = flow_mac_bare_metal_network.foobar.id
+}
+
+resource "flow_mac_bare_metal_security_group_rule" "foobar" {
+	security_group_id = flow_mac_bare_metal_security_group.foobar.id
+
+	direction = "ingress"
+	protocol  = { name = "icmp" }
+
+	icmp = {
+		type = %[2]d
+		code = %[3]d
+	}
+
+	ip_range = "%[4]s"
 }
 `